небольшие правки

+++ src/track/track.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, HttpException, HttpStatus } from '@nestjs/common';
 import { Track, Prisma } from '@prisma/client';
 import { SlugifyService } from '../slugify.service';
 import { PrismaService } from '../prisma.service';
@@ -23,33 +23,57 @@ export class TrackService {
     pictureFile && (picture = this.fileService.createFile(FileType.IMAGE, pictureFile));
     audioFile && (audio = this.fileService.createFile(FileType.AUDIO, audioFile));
     const slug: string = this.slugifyService.toSlug(name);
+    if (albumId) {
+      const albumCandidate = await this.prismaService.album.findUnique({ where: { id: albumId }, select: { id: true } });
+      if (!albumCandidate) throw new HttpException('Указанного альбома не существует в базе данных', HttpStatus.INTERNAL_SERVER_ERROR);
+    }
+    if (artistId) {
+      const artistCandidate = await this.prismaService.artist.findUnique({ where: { id: artistId }, select: { id: true } });
+      if (!artistCandidate) throw new HttpException('Указанного исполнителя не существует в базе данных', HttpStatus.INTERNAL_SERVER_ERROR);
+    }
     return this.prismaService.track.create({ data: { name, artistId, albumId, picture, audio, slug } });
   }
 
-  async update(data, id: number, files): Promise<Track> {
+  async update(
+    id: number,
+    name?: string, 
+    albumId?: number, 
+    artistId?: number, 
+    pictureFile?: Express.Multer.File, 
+    audioFile?: Express.Multer.File
+  ): Promise<Track> {
     const track = await this.prismaService.track.findUnique({ where: { id } });
     let picture: string, 
         audio: string, 
         slug: string;
 
-    if (data?.name) { // если меняем имя то меняем и slug
-      slug = this.slugifyService.toSlug(data?.name);
+    if (name) { // если меняем имя то меняем и slug
+      slug = this.slugifyService.toSlug(name);
     }
     
-    if (files?.picture) {
+    if (pictureFile) {
       // удалить старый файл
-
+      this.fileService.removeFile(track.picture);
       // записать новый
-      picture = this.fileService.createFile(FileType.IMAGE, files.picture[0]);
+      picture = this.fileService.createFile(FileType.IMAGE, pictureFile);
     }
 
-    if (files?.audio) {
+    if (audioFile) {
       // удалить старый файл
+      this.fileService.removeFile(track.audio);
       // записать новый
-      audio = this.fileService.createFile(FileType.AUDIO, files.audio[0]);
+      audio = this.fileService.createFile(FileType.AUDIO, audioFile);
     }
 
-    return this.prismaService.track.update({ data: { ...data, slug, picture, audio }, where: { id } });
+    if (albumId) {
+      const albumCandidate = await this.prismaService.album.findUnique({ where: { id: albumId }, select: { id: true } });
+      if (!albumCandidate) throw new HttpException('Указанного альбома не существует в базе данных', HttpStatus.INTERNAL_SERVER_ERROR);
+    }
+    if (artistId) {
+      const artistCandidate = await this.prismaService.artist.findUnique({ where: { id: artistId }, select: { id: true } });
+      if (!artistCandidate) throw new HttpException('Указанного исполнителя не существует в базе данных', HttpStatus.INTERNAL_SERVER_ERROR);
+    }
+    return this.prismaService.track.update({ data: { name, albumId, artistId, slug, picture, audio }, where: { id } });
   }
 
   async getAll(count: string, offset: string): Promise<Track[]> {
@@ -61,9 +85,13 @@ export class TrackService {
   }
 
   async delete(where: Prisma.TrackWhereUniqueInput): Promise<void> {
-    const track = await this.prismaService.track.delete({ where });
-    this.fileService.removeFile(track.audio);
-    this.fileService.removeFile(track.picture);
+    if (await this.prismaService.track.findUnique({ where })) {
+      const track = await this.prismaService.track.delete({ where });
+      this.fileService.removeFile(track.audio);
+      this.fileService.removeFile(track.picture);
+    } else {
+      throw new HttpException('Указанной композиции не существует', HttpStatus.INTERNAL_SERVER_ERROR);
+    }
   }
 
   async listen(id: number): Promise<Track> {
